+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/api/statement-commands.kt
@@ -25,5 +25,6 @@ interface UpdateStatementUseCase {
 
 interface DeleteStatementUseCase {
     fun remove(statementId: StatementId)
+    fun remove(statementIds: Set<StatementId>)
     fun removeAll()
 }